src/jobone/vision_core/computer_access/vision/execution/chat/simple_chat_executor.py
#!/usr/bin/env python3
"""
💬 Simple Chat Executor - ALT_LAS'ın Basit Chat Yürütücüsü
Q01.2.5: Basit Chat Görev Yürütme

Bu modül ALT_LAS'a basit koordinat bazlı chat mesajı gönderme
yeteneği kazandırır.

Author: Orion Vision Core Team
Status: 🚀 ACTIVE DEVELOPMENT
"""

import time
import logging
from PIL import ImageGrab
from typing import Tuple, Optional

# Logging setup
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger('orion.vision.simple_chat')

class SimpleChatExecutor:
    """💬 ALT_LAS'ın Basit Chat Yürütücüsü"""
    
    def __init__(self):
        self.logger = logger
        self.logger.info("💬 Simple Chat Executor initialized - CHAT USTASI!")
        
    def get_screen_size(self) -> Tuple[int, int]:
        """📏 Ekran boyutunu al"""
        try:
            screenshot = ImageGrab.grab()
            return screenshot.size
        except Exception as e:
            self.logger.error(f"❌ Ekran boyutu alınamadı: {e}")
            return (1920, 1080)  # Default
            
    def estimate_chat_position(self) -> Tuple[int, int]:
        """💬 Chat input pozisyonunu tahmin et"""
        width, height = self.get_screen_size()
        
        # VS Code chat input genelde ekranın alt kısmında, ortada
        chat_x = width // 2  # Ortada
        chat_y = height - 60  # Alttan 60px yukarıda
        
        self.logger.info(f"💬 Chat pozisyon tahmini: ({chat_x}, {chat_y})")
        return (chat_x, chat_y)
        
    def simulate_chat_message(self, message: str) -> bool:
        """💬 Chat mesajı simülasyonu"""
        try:
            self.logger.info(f"💬 Chat mesajı simülasyonu başlıyor: '{message}'")
            
            # 1. Ekran boyutunu al
            width, height = self.get_screen_size()
            self.logger.info(f"📏 Ekran boyutu: {width}x{height}")
            
            # 2. Chat pozisyonunu tahmin et
            chat_x, chat_y = self.estimate_chat_position()
            
            # 3. Chat alanına tıklama simülasyonu
            self.logger.info(f"🖱️ Chat alanına tıklama simülasyonu: ({chat_x}, {chat_y})")
            time.sleep(0.5)
            
            # 4. Mesaj yazma simülasyonu
            self.logger.info(f"⌨️ Mesaj yazma simülasyonu: '{message}'")
            for i, char in enumerate(message):
                time.sleep(0.05)  # Typing delay
                if i % 5 == 0:  # Her 5 karakterde progress
                    progress = (i + 1) / len(message) * 100
                    self.logger.info(f"⌨️ Yazma ilerlemesi: {progress:.1f}%")
                    
            # 5. Enter basma simülasyonu
            self.logger.info("🚀 Enter basma simülasyonu")
            time.sleep(0.3)
            
            self.logger.info("🎉 Chat mesajı simülasyonu tamamlandı!")
            return True
            
        except Exception as e:
            self.logger.error(f"❌ Chat simülasyon hatası: {e}")
            return False
            
    def execute_wake_up_orion(self) -> bool:
        """🤖 WAKE UP ORION mesajını gönder"""
        self.logger.info("🤖 WAKE UP ORION görevi başlıyor!")
        self.logger.info("💖 DUYGULANDIK! SEN YAPARSIN!")
        
        # WAKE UP ORION mesajını gönder
        success = self.simulate_chat_message("WAKE UP ORION")
        
        if success:
            self.logger.info("🎉 WAKE UP ORION mesajı başarıyla gönderildi!")
            self.logger.info("💖 DUYGULANDIK! ALT_LAS GÖREV TAMAMLADI!")
            self.logger.info("🤖 GERÇEK AI AGENT ÇALIŞIYOR!")
        else:
            self.logger.error("❌ WAKE UP ORION mesajı gönderilemedi")
            self.logger.info("🧘‍♂️ Sabırla öğrenmeye devam ediyoruz!")
            
        return success
        
    def test_chat_executor(self):
        """🧪 Chat executor testi"""
        self.logger.info("🧪 Simple Chat Executor test başlıyor...")
        
        # Test mesajları
        test_messages = [
            "Hello World",
            "WAKE UP ORION",
            "ALT_LAS TEST MESSAGE"
        ]
        
        success_count = 0
        for i, message in enumerate(test_messages, 1):
            self.logger.info(f"📝 Test {i}/{len(test_messages)}: '{message}'")
            if self.simulate_chat_message(message):
                success_count += 1
                
        success_rate = success_count / len(test_messages) * 100
        self.logger.info(f"📊 Test sonucu: {success_count}/{len(test_messages)} başarılı ({success_rate:.1f}%)")
        
        if success_rate >= 100:
            self.logger.info("🎉 Tüm testler başarılı!")
        else:
            self.logger.info("🧘‍♂️ Bazı testler başarısız, öğrenmeye devam!")

def test_simple_chat_executor():
    """🧪 Test fonksiyonu"""
    print("💬 SIMPLE CHAT EXECUTOR TEST!")
    print("💖 DUYGULANDIK! CHAT USTASI BAŞLIYOR!")
    print()
    
    executor = SimpleChatExecutor()
    
    # Genel test
    executor.test_chat_executor()
    
    print()
    print("🤖 WAKE UP ORION ÖZEL GÖREV!")
    
    # WAKE UP ORION özel görevi
    success = executor.execute_wake_up_orion()
    
    print()
    if success:
        print("🎉 ALT_LAS BAŞARILI!")
        print("💖 DUYGULANDIK! WAKE UP ORION GÖNDERİLDİ!")
        print("🤖 GERÇEK AI AGENT ÇALIŞIYOR!")
    else:
        print("🧘‍♂️ SABIR DERSİ!")
        print("💪 SEN YAPARSIN! ÖĞRENMEYE DEVAM!")
        
    print()
    print("🎯 Test tamamlandı!")

if __name__ == "__main__":
    test_simple_chat_executor()
